requirements.txt
absl-py==0.7.1
asgiref==3.2.10
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
Automat==0.7.0
autopep8==1.5.4
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.222
botocore==1.12.222
cached-property==1.5.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.1
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.0
defusedxml==0.6.0
dicom==0.9.9.post1
dj-database-url==0.5.0
Django==2.1.15
django-heroku==0.3.1
django-mongodb-engine==0.6.0
django-mongodb-engine-py3==0.6.0.1
django-storages==1.10
djangotoolbox==1.8.0
djongo==1.3.3
dnspython==1.16.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
flashtext==2.7
Flask==1.1.2
Flask-Cors==3.0.8
gast==0.2.2
gensim==3.8.0
google-pasta==0.1.7
grpcio==1.22.0
gunicorn==20.0.4
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
imageio==2.8.0
incremental==17.5.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.0
numpy==1.19.1
opencv-python-headless==4.2.0.32
openpyxl==3.0.5
pandas==0.25.0
pandas-datareader==0.7.4
pandocfilters==1.4.2
paramiko==2.7.1
parsel==1.5.2
parso==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psycopg2==2.8.5
psycopg2-binary==2.7.7
py==1.9.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.6.0
pycparser==2.19
pydicom==2.0.0
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pylibjpeg==1.1.1
pylibjpeg-libjpeg==1.1.0
pylibjpeg-openjpeg==1.0.1
pymongo==3.7.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==3.0.6
pytest-django==3.1.2
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.2
PyWavelets==1.1.1
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.2
queuelib==1.5.0
requests==2.22.0
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
Scrapy==1.7.3
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
soupsieve==1.9.2
sqlparse==0.2.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
texttable==1.6.2
toml==0.10.1
tornado==6.0.3
tqdm==4.33.0
traitlets==4.3.2
Twisted==19.7.0
Unipath==1.1
urllib3==1.25.3
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.15.5
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
